Beacon/Production/pyproject.toml
[project]
name = "production"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "btmgmt>=1.1.0",
    "paho-mqtt>=2.1.0",
]
